+++ NEWS.md
@@ -1,4 +1,9 @@
 
+# wrapr 1.9.6 2020-01-26
+
+ * Add dotpipe_eager_eval_*.
+ * Add on unpack method.
+
 # wrapr 1.9.5 2020-01-18
 
  * remove q= from sQuote to, as that feature wasn't added until newer Rs.